requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
Jinja2>=3.1,<4.0
PyYAML>=6.0,<7.0

aiogram==3.18.0

google-api-python-client>=2.120,<3.0
google-auth>=2.22,<3.0
google-auth-httplib2>=0.1.1,<1.0
google-auth-oauthlib>=1.2,<2.0

python-dotenv>=1.0,<2.0
